added function update

+++ app/src/main/java/reversetowerdefense/entities/Bird.java
@@ -2,7 +2,7 @@ package reversetowerdefense.entities;
 
 public class Bird extends Troop{
 
-    public Bird(String name, int hp, int speed, int damage, TroopType troopType, int hitSpeed) {
+    public Bird(String name, int hp, int speed, int damage, TroopType troopType, float hitSpeed) {
         super("Bird", 100, 1 , 5, TroopType.BIRD, 1);
 
     }
