+++ app/src/main/java/reversetowerdefense/entities/KingTower.java
@@ -2,7 +2,7 @@ package reversetowerdefense.entities;
 
 public class KingTower extends Tower{
 
-    public KingTower(String name, int hp, int range, TowerType towerType, TroopType spawnedTroop, int hitSpeed) {
+    public KingTower(String name, int hp, int range, TowerType towerType, TroopType spawnedTroop, float hitSpeed) {
         super("KING", 1000, 1, TowerType.KING, TroopType.SOLDIER,1);
         
     }
