+++ app/src/main/java/reversetowerdefense/entities/Ranger.java
@@ -2,7 +2,7 @@ package reversetowerdefense.entities;
 
 public class Ranger extends Troop{
 
-    public Ranger(String name, int hp, int speed, int damage, TroopType troopType, int hitSpeed) {
+    public Ranger(String name, int hp, int speed, int damage, TroopType troopType, float hitSpeed) {
         super("Ranger", 100, 1 , 5, TroopType.RANGER, 1);
 
     }
