+++ app/src/main/java/reversetowerdefense/entities/Soldier.java
@@ -2,7 +2,7 @@ package reversetowerdefense.entities;
 
 public class Soldier extends Troop{
 
-    public Soldier(String name, int hp, int speed, int damage, TroopType troopType, int hitSpeed) {
+    public Soldier(String name, int hp, int speed, int damage, TroopType troopType, float hitSpeed) {
         super("Soldier", 100, 1 , 5, TroopType.SOLDIER, 1);
         
     }
