+++ app/src/main/java/reversetowerdefense/entities/Tank.java
@@ -2,7 +2,7 @@ package reversetowerdefense.entities;
 
 public class Tank extends Troop{
 
-    public Tank(String name, int hp, int speed, int damage, TroopType troopType, int hitSpeed) {
+    public Tank(String name, int hp, int speed, int damage, TroopType troopType, float hitSpeed) {
         super("Tank", 100, 1 , 5, TroopType.TANK, 1);
 
     }
